refactor: move viam transport code out of dial
